+++ src/bewegung/core/vector/array3d.py
@@ -260,6 +260,14 @@ class VectorArray3D(VectorArray3DABC):
         return self._x.dtype
 
     @property
+    def ndim(self) -> int:
+        """
+        Number of dimensions
+        """
+
+        return 3
+
+    @property
     def mag(self) -> np.ndarray:
         """
         The vectors' magnitudes, computed on demand